+++ modules/mgr/chatmgr.js
@@ -1,4 +1,102 @@
 import { HMTABLES, HMCONST } from '../sys/constants.js';
+/* global PoolTerm */
+
+function getDiceSum(roll) {
+    let sum = 0;
+    for (let i = 0; i < roll.terms.length; i++) {
+        for (let j = 0; j < roll.terms[i]?.results?.length; j++) {
+            sum += roll.terms[i].results[j].result;
+        }
+    }
+    return sum;
+}
+
+async function saveExtendedTrauma(content, roll) {
+    const rolls = [roll];
+    let exContent = content;
+    const context = {
+        'duration': Math.max(roll.total * 5, 0),
+        'special': 'HM.passed',
+        'forever': false,
+    };
+
+    if (roll.total > 0) {
+        context.unit    = 'HM.seconds';
+        context.status  = 'HM.incapacitated';
+        context.special = 'HM.failed';
+
+        if (getDiceSum(roll) === 20) {
+            let newroll = await new Roll('5d6p').evaluate({async: true});
+            rolls.push(newroll);
+            let flavor = `${game.i18n.localize('HM.knockout')} ${game.i18n.localize('HM.duration')}`;
+            exContent += `<br> ${await newroll.render({flavor})}`;
+            context.duration = newroll.total;
+            context.unit     = 'HM.minutes';
+            context.status   = 'HM.knockedout';
+            context.special  = 'HM.critical';
+
+            newroll = await new Roll('d20').evaluate({async: true});
+            rolls.push(newroll);
+            flavor = `${game.i18n.localize('HM.comatose')} ${game.i18n.localize('HM.check')}`;
+            exContent += `<br> ${await newroll.render({flavor})}`;
+
+            if (newroll.total === 20) {
+                context.duration = newroll.total;
+                context.unit     = 'HM.days';
+                context.status   = 'HM.comatose';
+                context.special  = 'HM.doublecritical';
+
+                newroll = await new Roll('d20').evaluate({async: true});
+                rolls.push(newroll);
+                flavor = `${game.i18n.localize('HM.comatose')} ${game.i18n.localize('HM.duration')}`;
+                exContent += `<br> ${await newroll.render({flavor})}`;
+                if (newroll.total === 20) {
+                    context.forever = true;
+                    context.special = 'HM.goodbye';
+                    context.unit    = 'HM.indefinitely';
+                }
+            }
+        }
+    }
+
+    const template = 'systems/hackmaster5e/templates/chat/trauma.hbs';
+    const label = await renderTemplate(template, context);
+    return {content: label + exContent, rolls};
+}
+
+async function createSaveCard(roll, dataset, dialogResp) {
+    const {rollMode} = dialogResp.resp;
+    let saveType = (dataset.formulaType === 'fos' || dataset.formulaType === 'foa') ? 'Feat of ' : '';
+    if (dataset.ability) {
+        saveType = game.i18n.localize(`HM.abilityLong.${dataset.ability.toLowerCase()}`);
+    } else {
+        saveType += game.i18n.localize(`HM.saves.${dataset.formulaType}`);
+    }
+    const flavor = `${saveType} ${game.i18n.localize('HM.save')}`;
+    let content = await roll.render({flavor});
+
+    if (dataset.formulaType === 'trauma') {
+        const extended = await saveExtendedTrauma(content, roll);
+        content = extended.content;
+        const pool = PoolTerm.fromRolls(extended.rolls);
+        roll = Roll.fromTerms([pool]);
+    }
+    return {content, roll, rollMode};
+}
+
+async function createAbilityCard(roll, dataset) {
+    const saveType = game.i18n.localize(`HM.abilityLong.${dataset.ability.toLowerCase()}`);
+    const flavor = `${saveType} ${game.i18n.localize('HM.check')}`;
+    const content = await roll.render({flavor});
+    return {content};
+}
+
+async function createToPAlert(dataset) {
+    const template = 'systems/hackmaster5e/templates/chat/top.hbs';
+    const content = await renderTemplate(template, dataset);
+    const flavor = dataset.context.parent.name;
+    return {content, flavor};
+}
 
 export class HMChatMgr {
     constructor() { this._user = game.user.id; }
@@ -20,12 +118,12 @@ export class HMChatMgr {
                     cData = await this._createSkillCard(roll, dataset, dialogResp);
                     break;
                 case 'save':
-                    cData = await this._createSaveCard(roll, dataset);
+                    cData = await createSaveCard(roll, dataset, dialogResp);
                     break;
                 case 'ability':
                     cData = dialogResp.resp.save
-                        ? await this._createSaveCard(roll, dataset)
-                        : await this._createAbilityCard(roll, dataset);
+                        ? await createSaveCard(roll, dataset, dialogResp)
+                        : await createAbilityCard(roll, dataset);
                     break;
                 default:
             }
@@ -37,27 +135,17 @@ export class HMChatMgr {
             user:    this._user,
             flavor:  cData?.flavor || dialogResp?.caller?.name,
             content: cData.content,
-            type:    CONST.CHAT_MESSAGE_TYPES.OTHER,
+            type:    CONST.CHAT_MESSAGE_TYPES.IC,
         };
 
         if (roll) {
-            chatData.roll     = roll;
+            chatData.roll     = cData?.roll || roll;
             chatData.rollMode = cData.rollMode ? cData.rollMode : game.settings.get('core', 'rollMode');
             chatData.type     = CONST.CHAT_MESSAGE_TYPES.ROLL;
             chatData.sound    = CONFIG.sounds.dice;
         }
 
         return {...chatData, ...options};
-    }
-
-    getDiceSum(roll) {
-        let sum = 0;
-        for (let i = 0; i < roll.terms.length; i++) {
-            for (let j = 0; j < roll.terms[i]?.results?.length; j++) {
-                sum += roll.terms[i].results[j].result;
-            }
-        }
-        return sum;
     }
 
     async _createWeaponCard(roll, dataset, dialogResp) {
@@ -79,7 +167,7 @@ export class HMChatMgr {
                                 <b>${game.i18n.localize(`HM.${dialogResp.resp.rangestr}`)}</b>`;
 
                 let specialRow = '';
-                const sumDice = this.getDiceSum(roll);
+                const sumDice = getDiceSum(roll);
                 if (sumDice >=  20) { specialRow += '<b>Critical!</b>';        } else
                 if (sumDice === 19) { specialRow += '<b>Near Perfect</b>';     } else
                 if (sumDice === 1)  { specialRow += '<b>Potential Fumble</b>'; }
@@ -100,7 +188,7 @@ export class HMChatMgr {
                                 <b>${item.data.data.bonus.total.spd}</b>`;
 
                 let specialRow = '';
-                const sumDice = this.getDiceSum(roll);
+                const sumDice = getDiceSum(roll);
                 if (sumDice >=  20) { specialRow += '<b>Critical!</b>';        } else
                 if (sumDice === 19) { specialRow += '<b>Near Perfect</b>';     } else
                 if (sumDice === 1)  { specialRow += '<b>Potential Fumble</b>'; }
@@ -134,7 +222,7 @@ export class HMChatMgr {
                                 <b>${item.name}</b>`;
 
                 let specialRow = '';
-                const sumDice = this.getDiceSum(roll);
+                const sumDice = getDiceSum(roll);
                 if (sumDice >=  20) { specialRow += '<b>Perfect!</b>';     } else
                 if (sumDice === 19) { specialRow += '<b>Near Perfect</b>'; } else
                 if (sumDice === 18) { specialRow += '<b>Superior</b>';     } else
@@ -211,30 +299,4 @@ export class HMChatMgr {
         const content = await renderTemplate(template, dialogResp);
         return {content};
     }
-
-    async _createSaveCard(roll, dataset) {
-        let saveType = (dataset.formulaType === 'fos' || dataset.formulaType === 'foa') ? 'Feat of ' : '';
-        if (dataset.ability) {
-            saveType = game.i18n.localize(`HM.abilityLong.${dataset.ability.toLowerCase()}`);
-        } else {
-            saveType += game.i18n.localize(`HM.saves.${dataset.formulaType}`);
-        }
-        const flavor = `${saveType} ${game.i18n.localize('HM.save')}`;
-        const content = await roll.render({flavor});
-        return {content};
-    }
-
-    async _createAbilityCard(roll, dataset) {
-        const saveType = game.i18n.localize(`HM.abilityLong.${dataset.ability.toLowerCase()}`);
-        const flavor = `${saveType} ${game.i18n.localize('HM.check')}`;
-        const content = await roll.render({flavor});
-        return {content};
-    }
-}
-
-async function createToPAlert(dataset) {
-    const template = 'systems/hackmaster5e/templates/chat/top.hbs';
-    const content = await renderTemplate(template, dataset);
-    const flavor = dataset.context.parent.name;
-    return {content, flavor};
 }